+++ src/modules/sessions/repository/session.repository.ts
@@ -1,11 +1,11 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { CreateSessionDto } from '../dto/request/create.session.dto';
+import { CreateSessionDto } from '../dto/request/create.session.request';
 import { SessionEntity } from '../entities/session.entity';
 import { PrismaService } from '../../prisma/prisma.service';
-import { UpdateSessionDto } from '../dto/request/update.session.dto';
-import { PaginationQueryDto } from '../dto/request/pagination.query.dto';
+import { UpdateSessionDto } from '../dto/request/update.session.request';
+import { PaginationQueryDto } from '../dto/request/pagination.query.request';
 import { Prisma } from '@prisma/client';
-import { GetSessionsQueryDto } from '../dto/request/get.session.query.dto';
+import { GetSessionsQueryDto } from '../dto/request/get.session.query.request';
 
 @Injectable()
 export class SessionRepository {
