+++ src/modules/sessions/session.controller.ts
@@ -10,12 +10,12 @@ import {
     Query,
 } from '@nestjs/common';
 import { SessionService } from './session.service';
-import { CreateSessionDto } from './dto/request/create.session.dto';
+import { CreateSessionDto } from './dto/request/create.session.request';
 import { ApiOperation, ApiTags } from '@nestjs/swagger';
-import { GetSessionDto } from './dto/response/get.session.dto';
-import { UpdateSessionDto } from './dto/request/update.session.dto';
-import { PaginationQueryDto } from './dto/request/pagination.query.dto';
-import { GetSessionsQueryDto } from './dto/request/get.session.query.dto';
+import { GetSessionDto } from './dto/response/get.session.response';
+import { UpdateSessionDto } from './dto/request/update.session.request';
+import { PaginationQueryDto } from './dto/request/pagination.query.request';
+import { GetSessionsQueryDto } from './dto/request/get.session.query.request';
 
 @ApiTags('sessions')
 @Controller('/sessions')
