+++ src/modules/sessions/session.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@nestjs/common';
-import { CreateSessionDto } from './dto/request/create.session.dto';
+import { CreateSessionDto } from './dto/request/create.session.request';
 import { SessionRepository } from './repository/session.repository';
 import { SessionEntity } from './entities/session.entity';
-import { GetSessionDto } from './dto/response/get.session.dto';
-import { UpdateSessionDto } from './dto/request/update.session.dto';
-import { PaginationQueryDto } from './dto/request/pagination.query.dto';
-import { GetSessionsQueryDto } from './dto/request/get.session.query.dto';
+import { GetSessionDto } from './dto/response/get.session.response';
+import { UpdateSessionDto } from './dto/request/update.session.request';
+import { PaginationQueryDto } from './dto/request/pagination.query.request';
+import { GetSessionsQueryDto } from './dto/request/get.session.query.request';
 
 @Injectable()
 export class SessionService {
